Extract proposal-round MCP tools out of mcp-server.ts (Wave C batch 6)
mcp-server.ts 3976 -> 3656 lines. Splits open_proposal_round/
submit_proposal/get_proposal_status into mcp-proposal-tools.ts
(createProposalTools), along with their exclusive resolveProposerUid/
proposalRefusalHint helpers.

This cluster authenticates directly off the bearer (no job/session),
so it needs a different deps shape than prior batches: store,
agentTokenSecret, platformConnectorSecret, now, missingCredentialHint,
gamesStore, resolveProposalBase, contentChecker, onSourcesDelivered --
no resolveAuth/injectChannel.

matchesPlatformConnectorSecret + PLATFORM_CONNECTOR_ONLY_REASON move
into mcp-tool-support.ts since both this cluster and code staying in
mcp-server.ts (start, resolveAuth, create_game, open_round,
continue_draft) need them.
